defining http request
