+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node src/index.js",
+    "dev":"nodemon src/index.js"
   },
   "repository": {
     "type": "git",
@@ -17,8 +18,11 @@
   },
   "homepage": "https://github.com/tijaniabdellatif/Task-app-node#readme",
   "dependencies": {
+    "chalk": "^5.0.1",
+    "express": "^4.17.3",
     "mongodb": "^4.4.1",
     "mongoose": "^6.2.6",
+    "nodemon": "^2.0.15",
     "validator": "^13.7.0"
   }
 }
